refactor(api): uniform authentication handling
