+++ api/src/challenge/challenge.routes.ts
@@ -1,13 +1,15 @@
 import type express from 'express';
 import asyncHandler from 'express-async-handler';
 
-import { requestChallenge, verifyChallenge } from './challenge.controllers';
+import { requestChallenge, verifyChallenge } from './challenge.handlers';
+import { authBasic } from '../auth/auth.handlers';
 
 /**
  * Set up challenge routes.
  * @param app: express app instance.
  */
 export default (app: express.Express): void => {
+  app.use('/api/challenge/*', authBasic);
   app.post('/api/challenge/request', asyncHandler(requestChallenge));
   app.post('/api/challenge/verify', asyncHandler(verifyChallenge));
 };
