+++ api/src/friends/friends.routes.ts
@@ -13,12 +13,14 @@ import {
   getRequestsForFriend,
   getNames,
 } from './friends.handlers';
+import { authBasic } from '../auth/auth.handlers';
 
 /**
  * Set up friend routes.
  * @param app: express app instance.
  */
 export default (app: express.Express): void => {
+  app.use('/api/friends/*', authBasic);
   app.post('/api/friends/add', asyncHandler(addFriend));
   app.post('/api/friends/remove', asyncHandler(removeFriend));
   app.post('/api/friends/request', asyncHandler(requestAddFriend));
